pyproject.toml
[project]
name = "sophia-multimodal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "anthropic>=0.83.0",
    "chromadb>=1.5.1",
    "datasets>=4.5.0",
    "pillow>=12.1.1",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "transformers>=5.2.0",
]
